requirements_gpu.txt
scipy
numpy
jupyter
ipython
easydict
opencv-contrib-python==4.0.0.21
Cython
pillow
h5py
lmdb
mahotas
pandas
requests
# web.py==0.40.dev0
pika
matplotlib
lxml
bs4
tensorflow-gpu==1.13.1
keras
torch==1.0.0
torchvision==0.3.0
git+https://github.com/webpy/webpy.git#egg=webpy